Bug fixes and disposed of charts
Charts caused a big issue for merging the projects, took hours to debug, since the exception crashed the entire app without no sign, just a close. In the end, the only thing the exception said when we found a way to see it, it was that it doesn't recognize the exception text, not that much to go off. Andrei found this